+++ src/components/Card/myBusinessesCard.js
@@ -8,25 +8,20 @@ import { Edit } from '../Buttons/Edit'
 import "./CardComponent.css";
 
 export const MyBusinessCard = (props) => {
-    let name = props.name;
-    let website = props.website;
-    let owner = props.owner;
-    let type = props.type;
-    let description = props.description;
-    let address = props.address;
+    let info = props.info;
 
     return (
         <Card className='cardComponent'>
             <Row>
-                <Col><h2>Business name</h2></Col>
-                <Col className='right-col'><a href='' target='_blank'>Websitelink.com</a></Col>
+                <Col><h2>{info.name}</h2></Col>
+                <Col className='right-col'><a href='' target='_blank'>{info.businessDetails.websiteUrl}</a></Col>
             </Row>
             <Row>
-                <Col><p className='details'>Women owned</p></Col>
-                <Col className='right-col details'><p>Coffee</p></Col>
+                <Col><p className='details'>{info.ownerTypes[0].name} owned</p></Col>
+                <Col className='right-col details'><p>{info.businessType.name}</p></Col>
             </Row>
             <Row>
-                <Col lg={8} xs={10}><p>Description Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. </p></Col>
+                <Col lg={8} xs={10}><p>{info.businessDetails.description}</p></Col>
                 <Col></Col>
             </Row>
             <Row>
